Changed Validation Codes on Test Start, UI, and Internet Reconnect
Removed UpdateTest(8) in each test handler and instead just assigned the value of 8 in the InsertTest() function.

Changed UI to have predetermined test selection buttons (easy to change to different selections)

Uncommented the SQL reconnection code and added the internet reconnect via PHPWebRequest when checking the sql connection. So, the users on the settings menu and the logins should work even after sitting for a while prior to testing or after returning from the ending screen.